requirements.txt
requests
python-dotenv
pandas
jupyter
matplotlib
